backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
lancedb==0.20.0
google-generativeai==0.8.4
PyPDF2==3.0.1
python-docx==1.1.2
numpy>=1.24.0,<2.1
pydantic==2.10.4
pytest==8.3.4
httpx==0.28.1
